Add files via upload
Found a bug in the falling code which caused a random spike of delta time when a level loads. This could cause the player to fall at silly speeds and through the floor. This is resolved down to about 20fps by manually resetting delta time to 0 if it's too high